make DashboardItem.js more flexible clean up and remove DownloadItem.js
